refactor: `supported_devices` API を `tts_pipeline` router へ移動
